Check if cleaning up helps
